requirements.txt
numpy
pandas
matplotlib
opencv-contrib-python
tensorflow
keras
eventlet
Flask
python-socketio
